Refactored.
