+++ toolchain.common/Parsing/TokensIterator.cs
@@ -1,9 +1,15 @@
-﻿using System;
+﻿/////////////////////////////////////////////////////////////////////////////////////
+//
+// chibicc-toolchain - The specialized backend toolchain for chibicc-cil
+// Copyright (c) Kouji Matsui(@kozy_kekyo, @kekyo @mastodon.cloud)
+//
+// Licensed under MIT: https://opensource.org/licenses/MIT
+//
+/////////////////////////////////////////////////////////////////////////////////////
+
+using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.IO;
-using System.Linq;
-using chibicc.toolchain.Logging;
 using chibicc.toolchain.Tokenizing;
 
 namespace chibicc.toolchain.Parsing;
